+++ shutdown.sh
@@ -1,0 +1,4 @@
+/bin/echo -e "\nStopping LiveSectional" >> /var/log/livesectional.log
+/usr/bin/python /LiveSectional/live_sectional_daemon.py stop
+/usr/bin/python /LiveSectional/wipe.py
+exit