+++ startup.sh
@@ -1,4 +1,4 @@
-sudo echo -e "\nStarting up LiveSectional" >> /var/log/livesectional.log
-sudo python /LiveSectional/test.py
-#sudo python /LiveSectional/live_sectional.py
+/bin/echo -e "\nStarting up LiveSectional" >> /var/log/livesectional.log
+/usr/bin/python /LiveSectional/test.py
+/usr/bin/python /LiveSectional/live_sectional_daemon.py start
 exit